# libs/api-tosspaper/openapi.yaml
openapi: 3.1.0
servers:
  - url: http://localhost:8080
    description: Development server
  - url: http://localhost:8080
    description: Local development server
info:
  title: Purchase Order System API
  version: v1
  description: REST API for creating resources in the Purchase Order System.
  contact:
    name: Acme Corp
    url: https://localhost:8080
components:
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          example: "Acme Analytics Platform"
        message:
          type: string
          example: "An error occurred while processing your request"

    # Sender Approval Request
    ApproveSenderRequest:
      type: object
      required:
        - senderIdentifier
        - approvalStatus
      properties:
        senderIdentifier:
          type: string
          description: Email address (e.g., "user@example.com") or domain (e.g., "example.com") to approve/reject
          example: "supplier@acme.com"
        approvalStatus:
          type: string
          enum: [approved, rejected]
          description: |
            - approved: Approve sender and trigger extraction for all their attachments
            - rejected: Reject sender and soft-delete all their threads (scheduled for permanent deletion after grace period)

    # Extraction Review Request
...
    CompanyInfoUpdate:
      type: object
      x-class-extra-annotation: "@com.tosspaper.common.validator.AtLeastOneNotNull"
      properties:
        description:
          type: string
          nullable: true
          description: Short description
          example: "Leading provider of road-runner-proof products."
        assignedEmail:
          type: string
          format: email
          nullable: true
          description: Email selected for use   it has to be unique and not already assigned to another company
          example: "manager@useassetiq.com"
          x-field-extra-annotation: "@com.tosspaper.common.validator.ValidAssignedEmailDomain"
        logoUrl:
          type: string
          format: url
          nullable: true
          description: URL to the company logo
          example: "https://cdn.yourdomain.com/logos/company-42/logo-abc123.png"
          x-field-extra-annotation: "@com.tosspaper.common.validator.ValidUrl"
        termsUrl:
          type: string
          format: url
          nullable: true
          description: Link to terms and conditions
          example: "https://acme.com/terms"
          x-field-extra-annotation: "@com.tosspaper.common.validator.ValidUrl"
        privacyUrl:
          type: string
          format: url
          nullable: true
          description: Link to privacy policy
          example: "https://acme.com/privacy"
          x-field-extra-annotation: "@com.tosspaper.common.validator.ValidUrl"
        name:
          type: string
          nullable: true
          description: Name of the company
          example: "Acme Corporation Updated"
          x-jakarta-validation-annotations:
            - "@jakarta.validation.constraints.Size(min=1, message=\"Name cannot be empty if provided\")"
        countryOfIncorporation:
          type: string
          nullable: true
          description: "ISO 3166-1 alpha-2 country code (e.g., US, CA, GB) indicating where the company is incorporated. Used to determine default currency."
          example: "US"
          maxLength: 2
          minLength: 2
        currency:
          type: string
          nullable: true
          description: "ISO 4217 currency code (e.g., USD, CAD, EUR)"
          example: "USD"
        multicurrencyEnabled:
          type: boolean
          nullable: true
          description: "Whether multicurrency is enabled for this company"
          example: false
    Company:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the company
          example: 123456789
        created_at:
          type: string
          format: date-time
          description: Timestamp of creation
          example: "2025-06-03T02:48:43Z"
        updated_at:
          type: string
          format: date-time
          description: Timestamp of last update
          example: "2025-06-03T02:48:43Z"
        name:
          type: string
          description: Name of the company
          example: "Acme Corporation"
          x-jakarta-validation-annotations:
            - "@jakarta.validation.constraints.NotBlank"
        email:
          type: string
          format: email
          description: Contact email for the company
          example:
        assignedEmail:
          type: string
          format: email
          description: Email of the assigned account manager or contact (must be @useassetiq.com)
          example: "controla@useassetiq.com"
          x-field-extra-annotation: "@com.tosspaper.common.validator.ValidAssignedEmailDomain"
        description:
          type: string
          nullable: true
          description: Short description
          example: "Leading provider of road-runner-proof products."
        logoUrl:
          type: string
          format: url
          nullable: true
          description: URL to the company logo
          example: "https://cdn.yourdomain.com/logos/company-42/logo-abc123.png"
          x-field-extra-annotation: "@com.tosspaper.common.validator.ValidUrl"
        termsUrl:
          type: string
          format: url
          nullable: true
          description: Link to terms and conditions
          example: "https://acme.com/terms"
          x-field-extra-annotation: "@com.tosspaper.common.validator.ValidUrl"
        privacyUrl:
          type: string
          format: url
          nullable: true
          description: Link to privacy policy
          example: "https://acme.com/privacy"
          x-field-extra-annotation: "@com.tosspaper.common.validator.ValidUrl"
        currency:
          type: string
          nullable: true
          description: "ISO 4217 currency code (e.g., USD, CAD, EUR)"
          example: "USD"
        multicurrencyEnabled:
          type: boolean
          nullable: true
          description: "Whether multicurrency is enabled for this company"
          example: false
        countryOfIncorporation:
          type: string
          nullable: true
          description: "ISO 3166-1 alpha-2 country code (e.g., US, CA, GB) indicating where the company is incorporated. Used to determine default currency."
          example: "US"
          maxLength: 2
          minLength: 2
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          description: "Current page number (1-indexed)"
          example: 2
        pageSize:
          type: integer
          description: "Number of items per page"
          example: 20
        totalPages:
          type: integer
          description: "Total number of pages"
          example: 5
        totalItems:
          type: integer
          description: "Total number of items across all pages"
          example: 95
        cursor:
          type: string
          nullable: true
          description: "Cursor for next page (URL-safe base64 encoded created_at|id, generated by backend)"
          example: "MjAyNS0xMS0wMlQyMDozMDoyNi4wNzI1NjMrMDB8MDFLOUM0NzZSVkJTU1NSMllFMTM1VzRERFRX"
...
    ReceivedMessage:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the message
        from:
          type: string
          description: Sender email address
        to:
          type: string
          description: Recipient email address
        subject:
          type: string
          description: Email subject
        body_html:
          type: string
          nullable: true
          description: HTML content of the email body (if present)
        source:
          type: string
          description: Source of the message (email or direct_upload)
          enum: [email, direct_upload]
        date_received:
          type: string
          format: date-time
          description: When the message was received
        attachments_count:
          type: integer
          description: Number of attachments in this message
    Attachment:
      type: object
      properties:
        id:
          type: string
        fileName:
          type: string
        fileSize:
          type: integer
          format: int64
        status:
          type: string
        contentType:
          type: string
        storageUrl:
          type: string
          description: Full storage URL for the attachment
    ReceivedMessageList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ReceivedMessage'
        pagination:
          $ref: '#/components/schemas/Pagination'
    AttachmentList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
...
    ApprovedSenderResponse:
      type: object
      properties:
        id:
          type: string
          description: Approved sender ID
        companyId:
          type: integer
          format: int64
          description: Company ID
        senderIdentifier:
          type: string
          description: Email address or domain (e.g., user@example.com or example.com)
          example: "user@example.com"
        whitelistType:
          type: string
          enum: [email, domain]
          description: Type of whitelist - exact email or entire domain
        status:
          type: string
          enum: [approved, rejected]
          description: Approval status
        approvedBy:
          type: string
          description: User ID who approved this sender
        approvedAt:
          type: string
          format: date-time
          description: Timestamp when the sender was approved
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the sender was last updated
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the sender was created
        scheduledDeletionAt:
          type: string
          format: date-time
          nullable: true
          description: Timestamp when threads from this sender are scheduled for deletion (set when rejected, null when approved)
        domainAccessAllowed:
          type: boolean
          description: Whether domain-level approval/rejection is allowed for this sender (false if from public/disposable email domain like gmail.com, yahoo.com, etc.)
          example: false
    ApprovedSenderList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ApprovedSenderResponse'
        pagination:
          $ref: '#/components/schemas/Pagination'
    PendingApprovalDocument:
      type: object
      properties:
        attachmentId:
          type: string
          description: Attachment ID
        filename:
          type: string
          description: Filename of the attachment
        storageKey:
          type: string
          description: S3 storage key for the attachment
        messageId:
          type: string
          description: Email message ID
        dateReceived:
          type: string
          format: date-time
          description: Date when the email was received
    PendingSenderApproval:
      type: object
      properties:
        senderIdentifier:
          type: string
          description: Sender email address or domain
        documentsPending:
          type: integer
          description: Count of documents pending approval from this sender
        domainAccessAllowed:
          type: boolean
          description: Whether the sender's domain is allowed (false if disposable or personal email domain)
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/PendingApprovalDocument'
          description: List of attachments pending approval
    PendingSenderApprovalList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/PendingSenderApproval'
...
  responses:
    BadRequest:
      description: Bad Request - The request could not be understood by the server due to malformed syntax.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - Authentication credentials are required and have failed or have not yet been provided.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden - The server understood the request, but is refusing to fulfill it.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not Found - The requested resource could not be found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: Internal Server Error - A generic error message, given when an unexpected condition was encountered.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
security:
  - jwtAuth: []
paths:
...
  /v1/companies/{id}:
    get:
      operationId: getCompanyById
      tags:
        - Company
      summary: Get Company Information
      description: Retrieves information for a specific company.
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the company to retrieve
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved company information.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/InternalServerError'
    put:
      operationId: updateCompany
      tags:
        - Company
      summary: Update Company Information
      description: Updates attributes of an existing company.
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the company to update
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyInfoUpdate'
      responses:
        '204':
          description: Company updated successfully. No Content.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/InternalServerError'
...
  /v1/received_messages:
    get:
      operationId: listReceivedMessages
      tags:
        - "Received Messages"
      summary: "List Received Messages"
      description: "Retrieves a paginated list of received messages with optional date range filtering"
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
          description: Page number (1-indexed)
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            default: 20
          description: Number of items per page
        - name: status
          in: query
          required: false
          schema:
            type: string
          description: Filter by message status
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Search term (searches across from/to/cc/bcc/subject/body)
        - name: X-Context-Id
          in: header
          required: true
          schema:
            type: string
          description: Company ID
        - name: createdDateFrom
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Start date for filtering in UTC (ISO 8601 format, e.g., 2025-10-30T00:00:00Z)
        - name: createdDateTo
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: End date for filtering in UTC (ISO 8601 format, e.g., 2025-10-30T23:59:59Z)
        - name: fromEmail
          in: query
          required: false
          schema:
            type: string
          description: Filter by sender email address (exact match after cleaning)
      responses:
        "200":
          description: Successfully retrieved messages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceivedMessageList'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/InternalServerError'
  /v1/received_messages/{messageId}/attachments:
    get:
      operationId: getAttachments
      tags:
        - "Received Messages"
      summary: "Get Message Attachments"
      description: "Retrieves all attachments for a specific message"
      parameters:
        - name: messageId
          in: path
          required: true
          description: ID of the message
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Successfully retrieved attachments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttachmentList'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/InternalServerError'
...
  /v1/sender-approvals:
    get:
      operationId: listApprovedSenders
      tags:
        - "Sender Approvals"
      summary: "List approved/rejected senders"
      description: |
        Retrieve approved or rejected email senders for the authenticated company.
        Returns paginated list of sender records with their approval status.
      parameters:
        - name: X-Context-Id
          in: header
          required: true
          description: Company ID
          schema:
            type: string
        - name: status
          in: query
          required: false
          description: Filter by sender approval status (defaults to approved)
          schema:
            type: string
            enum: [approved, rejected]
            default: approved
        - name: page
          in: query
          required: false
          description: Page number (1-indexed)
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          required: false
          description: Number of items per page
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: "Paginated list of approved senders"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovedSenderList'
        "401":
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/InternalServerError'
    
    post:
      operationId: approveSender
      tags:
        - "Sender Approvals"
      summary: "Approve or reject a sender (email or domain)"
      description: |
        Approve or reject an email sender or entire domain.
        
        **Idempotency**: This endpoint is idempotent. Multiple calls with the same parameters will produce the same result.
        - If already approved, calling approve again has no side effects
        - If already rejected, calling approve will restore threads and trigger extraction
        
        **Approval:**
        - Email: Restores threads from that sender and triggers extraction for all their attachments
        - Domain: Restores ALL threads from that domain and triggers extraction for all attachments
        
        **Rejection:**
        - Email: Soft-deletes threads from that sender (scheduled for permanent deletion after grace period)
        - Domain: Soft-deletes ALL threads from that domain (cannot reject public domains like gmail.com)
        
        **Auto-Detection**: The system automatically detects whether the senderIdentifier is an email or domain.
      parameters:
        - name: X-Context-Id
          in: header
          required: true
          schema:
            type: string
          description: Company ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApproveSenderRequest'
      responses:
        "200":
          description: "Sender approval/rejection processed successfully. Extraction triggered if approved."
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          description: "Forbidden. Cannot approve/reject blocked domains (gmail.com, yahoo.com, etc.) or invalid sender format."
        default:
          $ref: '#/components/responses/InternalServerError'
  /v1/sender-approvals/pending-documents:
    get:
      operationId: listPendingDocuments
      tags:
        - "Sender Approvals"
      summary: "List pending documents grouped by sender"
      description: |
        Retrieve documents pending approval grouped by sender email.
        Shows sender email, count of pending documents, and list of attachments awaiting approval.
      parameters:
        - name: X-Context-Id
          in: header
          required: true
          description: Company ID
          schema:
            type: string
      responses:
        "200":
          description: "List of pending documents grouped by sender"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PendingSenderApprovalList'
        "401":
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/InternalServerError'

  # ============================================================
  # RBAC / Team Management Endpoints
  # ============================================================

  # User Invitations (no company context required)
  # Company Invitations (requires company context)
...